backend/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.30.1
pydantic>=2.7.4
supabase>=2.5.1
google-genai>=0.11.0 # Adjust to match local
python-dotenv>=1.0.1
pytest>=8.3.2
httpx>=0.27.0
